+++ src/ingestion/index.ts
@@ -11,6 +11,7 @@ export { smbcVpassAdapter } from './adapters/smbcVpass'
 export { securitiesAssetSnapshotAdapter } from './adapters/securitiesAssetSnapshot'
 export { sbiSecuritiesTradeHistoryAdapter } from './adapters/sbiSecuritiesTradeHistory'
 export { rakutenSecuritiesDomesticTradeHistoryAdapter } from './adapters/rakutenSecuritiesDomesticTradeHistory'
+export { monexUsStockTradeHistoryAdapter } from './adapters/monexUsStockTradeHistory'
 export { japaneseBrokerageTransactionsAdapter } from './adapters/japaneseBrokerageTransactions'
 export { moneyForwardAssetTrendAdapter } from './adapters/moneyForwardAssetTrend'
 export { moneyForwardHouseholdLedgerAdapter } from './adapters/moneyForwardHouseholdLedger'
@@ -26,6 +27,7 @@ import { smbcVpassAdapter } from './adapters/smbcVpass'
 import { securitiesAssetSnapshotAdapter } from './adapters/securitiesAssetSnapshot'
 import { sbiSecuritiesTradeHistoryAdapter } from './adapters/sbiSecuritiesTradeHistory'
 import { rakutenSecuritiesDomesticTradeHistoryAdapter } from './adapters/rakutenSecuritiesDomesticTradeHistory'
+import { monexUsStockTradeHistoryAdapter } from './adapters/monexUsStockTradeHistory'
 import { japaneseBrokerageTransactionsAdapter } from './adapters/japaneseBrokerageTransactions'
 import { moneyForwardAssetTrendAdapter } from './adapters/moneyForwardAssetTrend'
 import { moneyForwardHouseholdLedgerAdapter } from './adapters/moneyForwardHouseholdLedger'
@@ -42,6 +44,7 @@ export const importAdapters = [
   securitiesAssetSnapshotAdapter,
   sbiSecuritiesTradeHistoryAdapter,
   rakutenSecuritiesDomesticTradeHistoryAdapter,
+  monexUsStockTradeHistoryAdapter,
   japaneseBrokerageTransactionsAdapter,
   moneyForwardAssetTrendAdapter,
   moneyForwardHouseholdLedgerAdapter,